test(lang): move the cross-connection tests into query.spec.ts
